requirements.txt
evadb
openai
streamlit
python-environ
tabulate
eva-decord
